packages/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2
redis[hiredis]==5.0.8
celery[redis]==5.4.0
pydantic==2.8.2
pydantic-settings==2.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
pyotp==2.9.0
qrcode[pil]==7.4.2
httpx==0.27.0
boto3==1.34.144
Pillow==10.4.0
imagehash==4.3.1
structlog==24.2.0
prometheus-fastapi-instrumentator==7.0.0
python-multipart==0.0.9
slowapi==0.1.9
openpyxl==3.1.5
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
